+++ src/main/java/mate/academy/controllers/cars/AddDriverToCarController.java
@@ -1,4 +1,4 @@
-package mate.academy.controllers;
+package mate.academy.controllers.cars;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
