+++ src/main/java/mate/academy/controllers/cars/DisplayAllCarsController.java
@@ -1,23 +1,25 @@
-package mate.academy.controllers;
+package mate.academy.controllers.cars;
 
 import java.io.IOException;
+import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.lib.Injector;
-import mate.academy.service.DriverService;
+import mate.academy.model.Car;
+import mate.academy.service.CarService;
 
-public class DeleteDriverController extends HttpServlet {
+public class DisplayAllCarsController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("mate.academy");
-    private final DriverService driverService = (DriverService) injector
-            .getInstance(DriverService.class);
+    private final CarService carService = (CarService) injector.getInstance(CarService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        Long id = Long.valueOf(req.getParameter("id"));
-        driverService.delete(id);
-        resp.sendRedirect(req.getContextPath() + "/drivers/");
+
+        List<Car> cars = carService.getAll();
+        req.setAttribute("cars", cars);
+        req.getRequestDispatcher("/WEB-INF/views/cars/cars.jsp").forward(req, resp);
     }
 }
