+++ src/main/java/mate/academy/controllers/drivers/DisplayAllDriversController.java
@@ -1,4 +1,4 @@
-package mate.academy.controllers;
+package mate.academy.controllers.drivers;
 
 import java.io.IOException;
 import java.util.List;
