+++ src/main/java/mate/academy/controllers/manufacturer/CreateManufacturerController.java
@@ -1,4 +1,4 @@
-package mate.academy.controllers;
+package mate.academy.controllers.manufacturer;
 
 import java.io.IOException;
 import javax.servlet.ServletException;
