+++ src/main/java/mate/academy/controllers/manufacturer/DeleteManufacturerController.java
@@ -1,16 +1,14 @@
-package mate.academy.controllers;
+package mate.academy.controllers.manufacturer;
 
 import java.io.IOException;
-import java.util.List;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import mate.academy.lib.Injector;
-import mate.academy.model.Manufacturer;
 import mate.academy.service.ManufacturerService;
 
-public class DisplayAllManufacturersController extends HttpServlet {
+public class DeleteManufacturerController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("mate.academy");
     private final ManufacturerService manufacturerService = (ManufacturerService) injector
             .getInstance(ManufacturerService.class);
@@ -18,10 +16,8 @@ public class DisplayAllManufacturersController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-
-        List<Manufacturer> manufacturers = manufacturerService.getAll();
-        req.setAttribute("manufacturers", manufacturers);
-        req.getRequestDispatcher("/WEB-INF/views/manufacturers/manufacturers.jsp")
-                .forward(req, resp);
+        Long id = Long.valueOf(req.getParameter("id"));
+        manufacturerService.delete(id);
+        resp.sendRedirect(req.getContextPath() + "/manufacturers/");
     }
 }
